+++ cli/src/commands/git/push.rs
@@ -191,6 +191,9 @@ pub struct GitPushArgs {
     /// Only display what will change on the remote
     #[arg(long)]
     dry_run: bool,
+    /// Git push options
+    #[arg(long, short)]
+    option: Vec<String>,
 }
 
 fn make_bookmark_term(bookmark_names: &[impl fmt::Display]) -> String {
@@ -446,7 +449,14 @@ pub fn cmd_git_push(
     };
     let git_settings = tx.settings().git_settings()?;
     let push_stats = with_remote_git_callbacks(ui, |cb| {
-        git::push_branches(tx.repo_mut(), &git_settings, remote, &targets, cb)
+        git::push_branches(
+            tx.repo_mut(),
+            &git_settings,
+            remote,
+            &targets,
+            &args.option.iter().map(|s| s.as_str()).collect::<Vec<_>>(),
+            cb,
+        )
     })?;
     process_push_stats(&push_stats)?;
     tx.finish(ui, tx_description)?;