+++ CHANGELOG.md
@@ -1292,6 +1292,10 @@ Thanks to the people who made this release happen!
 
 ### New features
 
+* Add support for push options in `jj git push` with the `--option` flag.
+  This allows users to pass options to the remote server when pushing commits.
+  The short alias `-o` is also supported.
+
 * Add new boolean config knob, `ui.movement.edit` for controlling the behaviour
   of `prev/next`. The flag turns `edit` mode `on` and `off` permanently when set
   respectively to `true` or `false`.